[NuGet] Disable registry check on Mono.
Fixed bug #29088 - Custom NuGet feed package operations fail.
https://bugzilla.xamarin.com/show_bug.cgi?id=29088

A DisableBuffering option was added for the NuGet.exe push command in
b0e55ac46689248329b2be5b31e5b001ddc44ab3

This new feature checked the registry to see if .NET 4.5 is installed.
With Mono 4.0 this caused a 'No access to the given key' error when
viewing a custom NuGet package source such as Team City. The problem
is that Mono 4.0 does not create the directory:

/Library/Frameworks/Mono.framework/Versions/4.0.0/etc/mono/registry/LocalMachine

The registry/LocalMachine part is missing.

To workaround this problem the registry check in NuGet is now disabled
when on Mono. If Mono 4.0 is changed to create the
registry/LocalMachine when it is installed then this change will not
be needed.

Source code change:
https://github.com/mrward/nuget/commit/c8803d523747abab7f94f759f3841501dd4d31dd